@@ -1,5 +1,5 @@
 //= require jquery  
-//= require jquery_ujs  
+//= require jquery_ujs
 //= require_self  
 //= require_tree .  
   
