@@ -1,9 +1,5 @@
 # Be sure to restart your server when you modify this file.
 
-ActiveSupport::Inflector.inflections do |inflect|
-  inflect.irregular 'categorie', 'categories'
-end
-
 # Add new inflection rules using the following format
 # (all these examples are active by default):
 # ActiveSupport::Inflector.inflections do |inflect|
@@ -12,3 +8,13 @@ end
 #   inflect.irregular 'person', 'people'
 #   inflect.uncountable %w( fish sheep )
 # end
+#
+# These inflection rules are supported but not enabled by default:
+# ActiveSupport::Inflector.inflections do |inflect|
+#   inflect.acronym 'RESTful'
+# end
+
+ActiveSupport::Inflector.inflections do |inflect|
+  # TODO: rename the categorie model to category
+  inflect.irregular 'categorie', 'categories'
+end
