@@ -4,4 +4,4 @@
 # If you change this key, all old signed cookies will become invalid!
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-Notes::Application.config.secret_token = 'd9022542876d20f2dc576e4bd4734829b61ac8deffe59bdb204a5eb6bf2e68d79d2adbff4eedebd9c4834d990a883ddafed6c108ab6a2d406e7f60208aa39074'
+Notes::Application.config.secret_token = '5b09e5c1743a941eefb06b52869add6e4b4550cb057165813455270e8348dcd1eb772d21c288e71e4cef84f51336c5c86a40fc5c7e8a96586c3a0dd489ce2884'
