@@ -11,7 +11,7 @@
 #
 # It's strongly recommended to check this file into your version control system.
 
-ActiveRecord::Schema.define(:version => 20111128110914) do
+ActiveRecord::Schema.define(:version => 20111125172136) do
 
   create_table "categories", :force => true do |t|
     t.string   "nom"
@@ -27,11 +27,6 @@ ActiveRecord::Schema.define(:version => 20111128110914) do
     t.datetime "updated_at"
     t.integer  "categorie_id"
     t.integer  "rang"
-  end
-
-  create_table "notes_notes", :force => true do |t|
-    t.integer "parent_id"
-    t.integer "child_id"
   end
 
   create_table "users", :force => true do |t|
